+++ src/components/PhotosGrid/index.tsx
@@ -1,9 +1,31 @@
 import React from 'react';
+import { ActivityIndicator } from 'react-native';
 import { createFragmentContainer, graphql } from 'react-relay';
 import styled from 'styled-components';
 
+const UploadingIndicator = () => {
+  return <UploadingIndicatorContainer>
+    <ActivityIndicator color="orange" />
+  </UploadingIndicatorContainer>
+}
 
-const PhotosGrid = ({photos, containerWidth}) => {
+const Photo = ({photo, resizeMode="cover", style}) => {
+  const {
+    mediumURL,
+
+    uploading=false,  // non graphql field, used in messenger when sending message
+  } = photo;
+
+  return (<PhotoImage
+    source={{uri: mediumURL}}
+    resizeMode={resizeMode}
+    style={style}
+  >
+    {uploading && <UploadingIndicator />}
+  </PhotoImage>);
+}
+
+export const PhotosGrid = ({photos, containerWidth}) => {
   const borderWidth = 4;
   const photosCount = photos.length;
 
@@ -23,9 +45,8 @@ const PhotosGrid = ({photos, containerWidth}) => {
       height = maxHeight;
       width = containerWidth;
     }
-    photosGrid = <PhotoImage
-      source={{uri: photo.mediumURL}}
-      resizeMode="cover"
+    photosGrid = <Photo
+      photo={photo}
       style={{width, height}}
     />
   } else if(photosCount === 2) {
@@ -33,14 +54,12 @@ const PhotosGrid = ({photos, containerWidth}) => {
     const width = Math.round((containerWidth - borderWidth) / 2);
     
     photosGrid = <Row>
-      <PhotoImage
-        source={{uri: photos[0].mediumURL}}
-        resizeMode="cover"
+      <Photo
+        photo={photos[0]}
         style={{width, height, marginRight: borderWidth}}
       />
-      <PhotoImage
-        source={{uri: photos[1].mediumURL}}
-        resizeMode="cover"
+      <Photo
+        photo={photos[1]}
         style={{width, height}}
       />
     </Row>
@@ -48,21 +67,18 @@ const PhotosGrid = ({photos, containerWidth}) => {
     const halfWidth = Math.round((containerWidth - borderWidth) / 2);
     photosGrid = <>
       <Row>
-        <PhotoImage
-          source={{uri: photos[0].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[0]}
           style={{width: containerWidth, height: 200}}
         />
       </Row>
       <Row style={{marginTop: borderWidth}}>
-        <PhotoImage
-          source={{uri: photos[1].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[1]}
           style={{width: halfWidth, height: halfWidth, marginRight: borderWidth}}
         />
-        <PhotoImage
-          source={{uri: photos[2].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[2]}
           style={{width:halfWidth, height: halfWidth}}
         />
       </Row>
@@ -71,26 +87,22 @@ const PhotosGrid = ({photos, containerWidth}) => {
     const thirdWidth = Math.round((containerWidth - borderWidth * 2) / 3);
     photosGrid = <>
       <Row>
-        <PhotoImage
-          source={{uri: photos[0].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[0]}
           style={{width: containerWidth, height: 250}}
         />
       </Row>
       <Row style={{marginTop: borderWidth}}>
-        <PhotoImage
-          source={{uri: photos[1].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[1]}
           style={{width: thirdWidth, height: thirdWidth, marginRight: borderWidth}}
         />
-        <PhotoImage
-          source={{uri: photos[2].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[2]}
           style={{width: thirdWidth, height: thirdWidth, marginRight: borderWidth}}
         />
-        <PhotoImage
-          source={{uri: photos[3].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[3]}
           style={{width:thirdWidth, height: thirdWidth}}
         />
       </Row>
@@ -101,26 +113,22 @@ const PhotosGrid = ({photos, containerWidth}) => {
     const restPhotosCount = photosCount - 5;
     photosGrid = <>
       <Row>
-        <PhotoImage
-          source={{uri: photos[0].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[0]}
           style={{width: halfWidth, height: halfWidth, marginRight: borderWidth}}
         />
-        <PhotoImage
-          source={{uri: photos[1].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[1]}
           style={{width: halfWidth, height: halfWidth}}
         />
       </Row>
       <Row style={{marginTop: borderWidth}}>
-        <PhotoImage
-          source={{uri: photos[2].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[2]}
           style={{width: thirdWidth, height: thirdWidth, marginRight: borderWidth}}
         />
-        <PhotoImage
-          source={{uri: photos[3].mediumURL}}
-          resizeMode="cover"
+        <Photo
+          photo={photos[3]}
           style={{width: thirdWidth, height: thirdWidth, marginRight: borderWidth}}
         />
         <PhotoImage
@@ -158,7 +166,10 @@ const Row = styled.View`
   flex-direction: row;
 `;
 
-const PhotoImage = styled.ImageBackground``;
+const PhotoImage = styled.ImageBackground`
+  align-items: center;
+  justify-content: center;
+`;
 
 const MorePhotos = styled.View`
   background-color: #000;
@@ -171,4 +182,14 @@ const MorePhotosLabel = styled.Text`
   color: #fff;
   font-weight: 500;
   font-size: 25px;
+`;
+
+const UploadingIndicatorContainer = styled.View`
+  width: 50px;
+  height: 50px;
+  background-color: rgba(0,0,0,0.8);
+  border-radius: 25px;
+  overflow: hidden;
+  align-items: center;
+  justify-content: center;
 `;