+++ src/screens/MessengerScreen/Chat/Layout.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Platform } from 'react-native';
 import styled from 'styled-components/native';
 
-import { Dialog, Message } from '../models';
+import { Dialog, Message, MessageGeneric, MessageQueueItem } from '../models';
 import MessageBox from './MessageBox';
 import MessagesList from './MessagesList';
 
@@ -15,14 +15,29 @@ export interface ComponentProps {
 }
 
 export default (props: ComponentProps) => {
+  const [ sendQueue, setSendQueue ] = useState<MessageQueueItem[]>([]);
+
+  const sendMessage = (msg: MessageQueueItem) => {
+    // TODO: тут нужно изюавиться от очереди
+    //       и добавить это в стор
+    setSendQueue([msg, ...sendQueue]);
+  }
+
+  const removeMessageFromQueue = (tempID: string) => {
+    setSendQueue(sendQueue.filter( msg => msg.id !== tempID));
+  }
+
   return <Container behavior={Platform.OS == "ios" ? "padding" : "height"} keyboardVerticalOffset={64}>
     <MessagesList
       messages={props.messages}
       dialog={props.dialog}
       senderKind={props.senderKind}
       senderID={props.senderID}
+
+      sendQueue={sendQueue as MessageGeneric[]}
+      removeMessageFromQueue={removeMessageFromQueue}
     />
-    <MessageBox />
+    <MessageBox sendMessage={sendMessage} dialog={props.dialog} />
   </Container>
 }