+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.3
+- send message - first implementation (work only on android)
+# 0.1.2
+- add messenger attachmenmt popup - first implementation
 # 0.1.1
 - add my developed apps
 - add games predefined section